+++ weblate/checks/tests/test_commands.py
@@ -22,12 +22,13 @@
 from io import StringIO
 
 from django.core.management import call_command
+from django.test import SimpleTestCase
 
 from weblate.trans.tests.test_commands import WeblateComponentCommandTestCase
 from weblate.trans.tests.test_models import RepoTestCase
 
 
-class PeriodicCommandTest(RepoTestCase):
+class ListSameCommandTest(RepoTestCase):
     def setUp(self):
         super().setUp()
         self.component = self.create_component()
@@ -41,3 +42,10 @@ class PeriodicCommandTest(RepoTestCase):
 class UpdateChecksTest(WeblateComponentCommandTestCase):
     command_name = "updatechecks"
     expected_string = "Processing"
+
+
+class ListTestCase(SimpleTestCase):
+    def test_list_checks(self):
+        output = StringIO()
+        call_command("list_checks", stdout=output)
+        self.assertIn(".. _check-same:", output.getvalue())